+++ HISTORY.md
@@ -8,3 +8,9 @@ Commands:
 * `devid`:  check the existence of BMC.
 * `fwinfo`: get build time and revision of the current firmware. 
 
+###v0.1.2
+Modify the functionality of `fwinfo` command.
+
+Commands: 
+* `devid`:  check the existence of BMC.
+* `fwinfo`: get revision, ODM customer, and suitable motherboard of the current firmware.